% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{branch_volume_byorder_qsm}
\alias{branch_volume_byorder_qsm}
\title{Branch volume by order QSM}
\usage{
branch_volume_byorder_qsm(cylinder)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM or RCT QSM that is returned by
\code{\link{read_tree_qsm}} or \code{\link{read_rct_qsm}} with remove_root
= TRUE.}
}
\value{
A data frame with the branching orders of the TreeQSM or RCT QSM and
their total volumes in liters.
}
\description{
Extracts the branch volumes by branching order from the cylinder data of a
TreeQSM or a RCT QSM (treeinfo needs to have been run with --branch_data to
access the required information).
}
\examples{
\dontrun{
# Read tree qsm and extract branch volumes by order
qsm <- read_rct_qsm(QSM_path = "path/to/qsm.mat")
branchvol_byorder <- branch_volume_byorder_qsm(cylinder = qsm$qsm1$cylinder)
}
}
