% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{calc_r}
\alias{calc_r}
\title{Distance from the center}
\usage{
calc_r(x, y, xc, yc)
}
\arguments{
\item{x}{X values of the points.}

\item{y}{Y values of the points.}

\item{xc}{X-coordinate of the center.}

\item{yc}{Y-coordinate of the center.}
}
\value{
The distance of 2D points to the center
}
\description{
Calculates the distance of each 2D point (X,Y) in a point cloud from the
center (xc, yc) of a circle.
}
\details{
Support function used to determine the DBH from a tree point cloud with
\code{\link{dbh_pc}} and \code{\link{dab_pc}}.
}
\examples{
\dontrun{
Ri <- calc_r(x_dbh,y_dbh,x_c,y_c)
R <- mean(Ri)
}
}
