% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{crown_area_alpha_qsm}
\alias{crown_area_alpha_qsm}
\title{Alpha crown area TreeQSM}
\usage{
crown_area_alpha_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The alpha crown area of the TreeQSM in square meters.
}
\description{
Extracts the alpha crown area from the treedata of a TreeQSM. This is
calculated during the TreeQSM process as the area (m$^2$) of the crown's
planar projection's alpha shape.
}
\examples{
\dontrun{
# Read tree qsm and extract tree height
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
caa <- crown_area_alpha_qsm(treedata = qsm$treedata)
}
}
