% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{crown_evenness_qsm}
\alias{crown_evenness_qsm}
\title{Crown evenness TreeQSM}
\usage{
crown_evenness_qsm(cylinder)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The crown evenness.
}
\description{
Calculates the crown evenness from a TreeQSM.
}
\details{
The crown evenness is defined as "The crown cylinders divided into 8 angular
bins. Ratio between the minimum heights of the highest and lowest bin."
(Akerblom et al., 2017 & Terryn et al., 2020). Crown cylinders are determined
with \code{\link{crownset_qsm}}.
}
\examples{
\dontrun{
QSM_path <- "path/to/qsm.mat"
qsm <- read_tree_qsm(QSM_path)
ce <- crown_evenness_qsm(qsm$cylinder, crown)
}
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
