% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{crown_start_height}
\alias{crown_start_height}
\title{Crown start height TreeQSM}
\usage{
crown_start_height(treedata, cylinder, crownset)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{crownset}{An integer containing the indices of the cylinders belonging
to the crownset which is returned by \code{\link{crownset}}.}
}
\value{
The crown start height.
}
\description{
Calculates the crown start height from a TreeQSM.
}
\details{
The crown start height is defined as "The height of the first stem branch in
the tree crown relative to the tree height" (Akerblom et al., 2017 & Terryn
et al., 2020).
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
qsm <- read_tree_qsm(QSM_path)
crown <- crownset(qsm$cylinder)
csh <- crown_start_height(qsm$treedata,qsm$cylinder,crown)
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
