% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{crown_start_height_qsm}
\alias{crown_start_height_qsm}
\title{Crown start height TreeQSM}
\usage{
crown_start_height_qsm(treedata, cylinder, pc = "NA")
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}. Default is "NA" and indicates no tree point
cloud is available.}
}
\value{
The crown start height.
}
\description{
Calculates the crown start height from a TreeQSM.
}
\details{
The crown start height is defined as "The height of the first stem branch in
the tree crown relative to the tree height" (Akerblom et al., 2017 & Terryn
et al., 2020). The tree height is calculated with \code{\link{tree_height}}.
Crown cylinders are determined with \code{\link{crownset_qsm}}.
}
\examples{
\dontrun{
QSM_path <- "path/to/qsm.mat"
qsm <- read_tree_qsm(QSM_path)
csh <- crown_start_height_qsm(qsm$treedata,qsm$cylinder)
PC_path <- "path/to/point_cloud.txt"
pc <- read_tree_pc(PC_path)
csh <- crown_start_height_qsm(qsm$treedata,qsm$cylinder,pc)
}
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
