% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{crownset}
\alias{crownset}
\title{Crownset TreeQSM}
\usage{
crownset(cylinder)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
An integer containing the indices of the cylinders belonging to the
crownset.
}
\description{
Returns the indices of the cylinders belonging to the crown of the treeQSM.
}
\details{
The crownset is determined based on four steps (that are designed to exclude
dead branches at the bottom of the stem) according to Akerblom et al. (2017).
STEP 1: Initialize the crown set as cylinders that have a branching order
higher than three. If the initial set is empty, the minimum order is lowered
until the set becomes non-empty. STEP 2: As long as the crown set extends,
append the parent cylinders of the crown set that are not part of the stem.
STEP 3: Append to the crown set cylinders that are not part of the stem but
whose start point is higher than the lowest starting point of crown cylinders
connected to the stem. STEP 4: As long as the crown set extends, append the
child cylinders of the crown set.
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
qsm <- read_tree_qsm(QSM_path)
crown <- crownset(qsm$cylinder)
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.
}
