% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{cylinder_length_volume_ratio_qsm}
\alias{cylinder_length_volume_ratio_qsm}
\title{Calculate the cylinder length volume ratio from a TreeQSM}
\usage{
cylinder_length_volume_ratio_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The cylinder length volume ratio defined as "the ratio between total
length and volume of the branch cylinders" (Akerblom et al., 2017 & Terryn et al., 2020).
}
\description{
Calculate the cylinder length volume ratio from a TreeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
len_vol_ratio <- cylinder_length_volume_ratio_qsm(treedata)
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M. (2017). Automatic tree species recognition with quantitative structure models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ... & Verbeeck, H. (2020). Tree species classification using structural features derived from terrestrial laser scanning. ISPRS Journal of Photogrammetry and Remote Sensing, 168, 170-181.
}
