% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{cylinder_volume_byradius_qsm}
\alias{cylinder_volume_byradius_qsm}
\title{Cylinder volume by radius size QSM}
\usage{
cylinder_volume_byradius_qsm(cylinder, radius_bin_size = 0.01)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM or RCT QSM that is returned by
\code{\link{read_tree_qsm}} or \code{\link{read_rct_qsm}} with remove_root
= TRUE.}

\item{radius_bin_size}{Numeric. The radius_bin_size determines the size that
the bins should be in metres. A value of 0.01 will divide the cylinders in
1 cm bins and count up the volumes of all cylinders in that size bin. 1 cm
is the default.}
}
\value{
A data frame with the radius ranges of the TreeQSM or RCT QSM and
their total volumes in liters.
}
\description{
Extracts the total cylinder volumes by radius size from the cylinder data of
a TreeQSM or a RCT QSM (treeinfo needs to have been run with --branch_data to
access the required information).
}
\examples{
\dontrun{
# Read tree qsm and extract branch volumes by order
qsm <- read_rct_qsm(QSM_path = "path/to/qsm.mat")
cylvol_byradius <- cylinder_volume_byradius_qsm(cylinder = qsm$qsm1$cylinder)
}
}
