% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{dab_pc}
\alias{dab_pc}
\title{Diameter above buttresses point cloud}
\usage{
dab_pc(pc, thresholdbuttress = 0.001, maxbuttressheight = 7, plot = FALSE)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{thresholdbuttress}{Numeric value (default=0.001) that is multiplied
with the radius to determine if the stem is circular or irregular at the
height the slice is taken. For example with the default value 0.001: when
the average residual (obtained after an initial circle fitting at 1.3 m)
exceeds a value of 0.001 times the radius, indicating a non-circular
(irregular) stem shape and presumably buttresses, the circle fitting
process is repeated with a new slice 6 mm higher than the previous one
until a slice above the buttresses is reached.}

\item{maxbuttressheight}{Numeric value (default=7) that limits the height at
which the diameter is measured. When this height is reached (because
residuals do not become smaller than thresholdbuttress * R), the
thresholdbuttress value is increased with 0.0005 and the fitting starts
again at 1.3 m.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
Diameter of the stem above buttresses (numeric value). Also
optionally (plot=TRUE) plots the circle fitting on the horizontal slice.
}
\description{
Returns the diameter above buttresses (DAB) of a tree measured from a tree
point cloud.
}
\details{
The DAB is measured as the diameter of the optimal circle fitted through a
6mm thick horizontal slice taken above the buttresses. A least squares circle
fitting algorithm was applied to find the optimal fit. The height at which
the horizontal slice is taken, is determined iteratively. Starting at 1.27 m
to 1.33 m from the lowest point of the tree point cloud. The average residual
between the points and the fitted circle is calculated. When the average
residual exceeds a value of "thresholdbuttress" times the radius, indicating
a non-circular (irregular) stem shape and presumably buttresses, the process
is repeated with a new slice 6 mm higher than the previous one until a slice
above the buttresses is reached. When the "maxbuttressheight" is exceeded the
iterative process is restarted with a "thresholdbuttress" increased with
0.0005.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the DAB
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
dab <- dab_pc(pc = pc_tree)
# and plot the circle fitting
dab <- dab_pc(pc = pc_tree, plot = TRUE)
# with non-default settings
dab <- dab_pc(pc = pc_tree, thresholdbuttress = 0.002, maxbuttressheight = 5,
              plot = TRUE)
}
}
