% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{dbh_height_ratio_qsm}
\alias{dbh_height_ratio_qsm}
\title{Calculate DBH-tree height ratio from a TreeQSM}
\usage{
dbh_height_ratio_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
DBH divided by the tree height.
}
\description{
Calculate DBH-tree height ratio from a TreeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
ratio_height <- dbh_height_ratio_qsm(treedata)
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M. (2017). Automatic tree species recognition with quantitative structure models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ... & Verbeeck, H. (2020). Tree species classification using structural features derived from terrestrial laser scanning. ISPRS Journal of Photogrammetry and Remote Sensing, 168, 170-181.
}
