% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{dbh_minradius_ratio}
\alias{dbh_minradius_ratio}
\title{DBH minimum tree radius ratio TreeQSM}
\usage{
dbh_minradius_ratio(treedata, cylinder)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The ratio of the dbh and the minimum tree radius.
}
\description{
Calculates the ratio between the dbh and the minimum tree radius from a
TreeQSM.
}
\details{
This ratio is defined as "Ratio between the DBH and the minimum of the
vertical bin radius estimates" (Akerblom et al., 2017 & Terryn et al., 2020).
The vertical bin radius estimates are calculated with
\code{\link{vertical_bin_radii}}.
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
qsm <- read_tree_qsm(QSM_path)
dbh_rad_ratio <- dbh_minradius_ratio(qsm$treedata,qsm$cylinder)
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
