% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{dbh_pc}
\alias{dbh_pc}
\title{Diameter at breast height point cloud}
\usage{
dbh_pc(pc, plot = FALSE)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
Diameter of the stem at breast height (numeric value). Also
optionally (plot=TRUE) plots the circle fitting on the horizontal slice.
}
\description{
Returns the diameter at breast height (DBH) of a tree measured from a tree
point cloud.
}
\details{
The DBH is measured as the diameter of the optimal circle fitted through a
6mm thick horizontal slice (from 1.27 m to 1.33 m above the lowest tree
point). A least squares circle fitting algorithm was applied to find the
optimal fit.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the DBH
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
dbh <- dbh_pc(pc = pc_tree)
# and plot the circle fitting
dbh <- dbh_pc(pc = pc_tree, plot = TRUE)
}
}
