% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_qsm}
\alias{example_qsm}
\title{Example qsm}
\format{
a list with four attributes (cylinder, branch, treedata, triangulation)
\describe{
\item{cylinder}{list of 13 cylinder attributes}
\item{branch}{list of 10 branch attributes}
\item{treedata}{list of 91 tree attributes}
\item{triangulation}{list of 12 triangulation attributes}
}
}
\source{
Louise Terryn (Louise.Terryn@UGent.be)
}
\usage{
example_qsm
}
\description{
An example TreeQSM (Raumonen et al., 2013, version 2.4.0).
}
\keyword{datasets}
