% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{f}
\alias{f}
\title{Algebraic distance from the center}
\usage{
f(c, x, y)
}
\arguments{
\item{c}{First estimate of the center coordinates to be optimised (xc,yc).}

\item{x}{X values of the points.}

\item{y}{Y values of the points.}
}
\value{
When optimised returns the optimised center estimate of the circle
fitting.
}
\description{
Calculates the algebraic distance between the data points and the mean circle
centered at c=(xc, yc) based on \code{\link{calc_r}}.
}
\details{
Support function used to determine the DBH from a tree point cloud with the
functions \code{\link{dbh_pc}} and \code{\link{dab_pc}}.
}
\examples{
#center_estimate <- optim(par=c(x_m,y_m), fn=f, x=x_dbh, y=y_dbh)
}
