% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapering_metrics.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Taper plot}
\usage{
lm_eqn(df)
}
\arguments{
\item{df}{Data.frame with diameter height information.}
}
\value{
Plot of the linear regression
}
\description{
Plots of the linear regression between diameter and height
}
\examples{
\dontrun{
taper_plot <- ggplot2::ggplot(T, ggplot2::aes(diameter, height)) +
ggplot2::geom_point() +
ggplot2::geom_smooth(method='lm') +
ggplot2::geom_text(x = (max(T$diameter)+min(T$diameter))/2,
                   y = max(T$height)-0.5, label = lm_eqn(T), parse = TRUE)
}
}
