% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavls_metrics_pc.R
\name{mean_tree_position_pc}
\alias{mean_tree_position_pc}
\title{Mean tree point cloud position}
\usage{
mean_tree_position_pc(pc)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}
}
\value{
Numeric with the XYZ coordinates (location) of the tree stem.
}
\description{
Returns the mean (X,Y,Z)-position of a tree point cloud.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the tree position
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
pos <- mean_tree_position_pc(pc = pc_tree)
}
}
