% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{normalize_pc}
\alias{normalize_pc}
\title{Normalize a tree point cloud}
\usage{
normalize_pc(pc, dtm = NA, r = 5)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{dtm}{The digital terrain model as a data.frame with columns X,Y,Z
(default = NA). If the digital terrain model is in the same format as a
point cloud it can also be read with \code{\link{read_tree_pc}}.}

\item{r}{Numeric value (default=5) r which determines the range taken for the
dtm. Should be at least the resolution of the dtm. Only relevant when a dtm
is provided.}
}
\value{
Normalized point cloud as a data.frame with columns X,Y,Z.
}
\description{
Normalizes a tree point cloud by subtracting from each column its respective
the min value (e.g. all X-values - min(all X-values)). When the bottom of the
point cloud is incomplete or obstructed you can choose to add a digital
terrain model as an input which is used to estimate lowest point of the point
cloud in order to obtain the correct normalized height of the tree.
}
\examples{
\dontrun{
# Read tree point cloud and normalise the point cloud
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
pc_norm <- normalize_pc(pc)
}
}
