% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_metrics_pc.R
\name{plot_cv_pcs}
\alias{plot_cv_pcs}
\title{Calculate and save figures of \code{\link{volume_crown_pc}} function}
\usage{
plot_cv_pcs(
  PCs_path,
  extension = ".txt",
  OUT_path,
  alpha = 1,
  thresholdbranch = 1.5,
  minheight = 1,
  buttress = FALSE,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7
)
}
\arguments{
\item{PCs_path}{A character with the path to the folder that contains the
tree point clouds.}

\item{extension}{A character refering to the file extension of the point
cloud files (default=".txt"). Can be ".txt", ".ply" or ".las".}

\item{OUT_path}{A character with the path to the folder where the figures
should be saved.}

\item{alpha}{Numeric value (default=1) alpha for the computation of the 3D
alpha-shape of the tree crown based on \code{\link[alphashape3d]{ashape3d}}
in \code{\link{volume_crown_pc}}.}

\item{thresholdbranch}{Numeric value (default=1.5) from
\code{\link{classify_crown_pc}}.}

\item{minheight}{Numeric value (default=1) from
\code{\link{classify_crown_pc}}. The default value is based on
non-buttressed trees. Choose a higher value (e.g. 4) for buttressed trees.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses (higher than breast height).}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses which is used in \code{\link{classify_crown_pc}}. Only relevant
when buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses which is used in \code{\link{classify_crown_pc}}. Only relevant
when buttress == TRUE.}
}
\value{
a numeric containing the crown volume values for each tree point
cloud. Figures are saved in the output folder.
}
\description{
Calculates the crown volume and saves the figures acquired when running
\code{\link{volume_crown_pc}} on multiple tree point clouds in a folder.
}
\details{
Uses \code{\link{read_tree_pc}} to read the point clouds and
\code{\link{volume_crown_pc}} with parameter plot = TRUE to calculate the
crown volume and plot the crown and 3D the alpha-shape fitting. For
buttressed trees, first optimise the thresholdbuttress, maxbuttressheight,
thresholdbranch and minheight parameter values using
\code{\link{plot_dab_fit_pcs}} and
\code{\link{plot_crown_classification_pcs}} and use those optimised values in
this function.
}
\examples{
\dontrun{
# Calculate CV with default settings and save alpha shape figures
cvs <- plot_cv_pcs(PCs_path = "path/to/folder/PCs/", extension = ".txt",
                   OUT_path = "path/to/figure/folder/")
# Calculate CV with non-default settings and save alpha shape figures
cvs <- plot_cv_pcs(PCs_path = "path/to/folder/PCs/", extension = ".txt",
                   OUT_path = "path/to/figure/folder/", alpha = 2)
# Calculate CV with non-default settings and save alpha shape figures
# for buttressed trees
cvs <- plot_cv_pcs(PCs_path = "path/to/folder/PCs/", extension = ".txt",
                   OUT_path = "path/to/figure/folder/", alpha = 2,
                   minheight = 4, buttress = TRUE)
}
}
