% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_metrics_pc.R
\name{plot_dab_fit_pcs}
\alias{plot_dab_fit_pcs}
\title{Calculate and save figures of \code{\link{dab_pc}} function}
\usage{
plot_dab_fit_pcs(
  PCs_path,
  extension = ".txt",
  OUT_path,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7
)
}
\arguments{
\item{PCs_path}{A character with the path to the folder that contains the
tree point clouds.}

\item{extension}{A character refering to the file extension of the point
cloud files (default=".txt"). Can be ".txt", ".ply" or ".las".}

\item{OUT_path}{A character with the path to the folder where the figures
should be saved.}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses.}
}
\value{
a numeric containing the dbh values for each tree point cloud.
Figures are saved in the output folder.
}
\description{
Calculates the dab and saves the figures acquired when running
\code{\link{dab_pc}} on multiple tree point clouds in a folder. Use different
values for the thresholdbuttress and maxbuttressheight parameter to optimise
dab calculation for your tree point clouds.
}
\details{
Uses \code{\link{read_tree_pc}} to read the point clouds and
\code{\link{dab_pc}} with parameter plot = TRUE to calculate the dab and plot
the circle fitting.
}
\examples{
\dontrun{
PCs_path <- "path/to/folder/PCs/"
extension <- ".txt"
OUT_path <- "path/to/folder/for/saving/figures/"
dab_values <- plot_dab_fit_pcs(PCs_path, extension, OUT_path)
}
}
