% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_metrics_pc.R
\name{plot_dbh_fit_pcs}
\alias{plot_dbh_fit_pcs}
\title{Calculate and save figures of \code{\link{dbh_pc}} function}
\usage{
plot_dbh_fit_pcs(PCs_path, extension = ".txt", OUT_path)
}
\arguments{
\item{PCs_path}{A character with the path to the folder that contains the
tree point clouds.}

\item{extension}{A character refering to the file extension of the point
cloud files (default=".txt"). Can be ".txt", ".ply" or ".las".}

\item{OUT_path}{A character with the path to the folder where the figures
should be saved.}
}
\value{
a numeric containing the dbh values for each tree point cloud.
Figures are saved in the output folder.
}
\description{
Calculates  the dbh and saves the figures acquired when running
\code{\link{dbh_pc}} on multiple tree point clouds in a folder.
}
\details{
Uses \code{\link{read_tree_pc}} to read the point clouds and
\code{\link{dbh_pc}} with parameter plot = TRUE to calculate the dbh and plot
the circle fitting.
}
\examples{
\dontrun{
# Calculate DBHs and save circle fitting figures
dbh_values <- plot_dbh_fit_pcs(PCs_path = "path/to/folder/PCs/",
                               extension = ".txt",
                               OUT_path = "path/to/figure/folder/")
}
}
