% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_metrics_pc.R
\name{plot_tree_height_pcs}
\alias{plot_tree_height_pcs}
\title{Calculate and save figures of \code{\link{tree_height_pc}} function}
\usage{
plot_tree_height_pcs(
  PCs_path,
  extension = ".txt",
  dtm = "NA",
  r = 5,
  OUT_path = "./"
)
}
\arguments{
\item{PCs_path}{A character with the path to the folder that contains the
tree point clouds.}

\item{extension}{A character refering to the file extension of the point
cloud files (default=".txt"). Can be ".txt", ".ply" or ".las".}

\item{dtm}{The digital terrain model (default = NA), parameter of
\code{\link{tree_height_pc}}.}

\item{r}{Numeric value (default=5) r, parameter of
\code{\link{tree_height_pc}}. Only relevant if a dtm is provided.}

\item{OUT_path}{A character with the path to the folder where the figures
should be saved (default = current folder).}
}
\value{
A list with in the first element a numeric containing the tree height
values for each tree point cloud. In the second element there is the list
with the plots. Figures are also saved in the output folder.
}
\description{
Calculates the tree height and saves the figures acquired when running
\code{\link{tree_height_pc}} on multiple tree point clouds in a folder.
}
\details{
Uses \code{\link{read_tree_pc}} to read the point clouds and
\code{\link{tree_height_pc}} with parameter plot = TRUE to calculate the tree
height and plot the tree point cloud.
}
\examples{
\dontrun{
# Calculate tree height and save figures
height_values <- plot_tree_height_pcs(PCs_path = "path/to/folder/PCs/",
                                  extension = ".txt",
                                  OUT_path = "path/to/figure/folder/")
# Calculate tree height using dtm of resolution 2 and save figures
dtm_df <- read_tree_pc("path/to/dtm.txt")
height_values <- plot_tree_height_pcs(PCs_path = "path/to/folder/PCs/",
                                  extension = ".txt", dtm = dtm_df, r = 2,
                                  OUT_path = "path/to/figure/folder/")
}
}
