% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{projected_area_pc}
\alias{projected_area_pc}
\title{Projected area point cloud}
\usage{
projected_area_pc(pc, concavity = 2, plot = FALSE)
}
\arguments{
\item{pc}{The point cloud as a data.frame with columns X,Y,Z (e.g. output of
\code{\link{read_tree_pc}}.}

\item{concavity}{Numeric value (default=2) concavity for the computation of a
concave hull based on \code{\link[concaveman]{concaveman}}.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
The projected area (numeric value) as the area of the concave hull
computed from the points of point cloud. Also optionally (plot=TRUE) plots
the concave hull fitting and in this case returns a list with the area as
first element and the plot as the second element.
}
\description{
Returns the projected area measured from a point cloud.
}
\details{
This function uses \code{\link[sf]{st_area}} and
\code{\link[concaveman]{concaveman}} to calculate the area of the concave
hull fitted to the provided point clouds.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the projected tree area
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
pta <- projected_crown_area_pc(pc = pc_tree)
# and plot the concave hull fitting
output <- projected_crown_area_pc(pc = pc_tree, plot = TRUE)
pca <- output$pca
# classify the tree point cloud and calculate the projected crown area
crown_pc <- classify_crown_pc(
  pc, thresholdbranch, minheight, buttress,
  thresholdR2, thresholdbuttress,
  maxbuttressheight, FALSE
)
pca <- projected_crown_area_pc(pc = crown_pc$crownpoints)
}
}
