% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{projected_crown_area_pc}
\alias{projected_crown_area_pc}
\title{Projected crown area point cloud}
\usage{
projected_crown_area_pc(
  pc,
  concavity = 2,
  thresholdbranch = 1.5,
  minheight = 4,
  plot = FALSE
)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{concavity}{Numeric value (default=2) concavity for the computation of a
concave hull based on \code{\link[concaveman]{concaveman}}.}

\item{thresholdbranch}{Numeric value (default=1.5) from
\code{\link{classify_crown_pc}}.}

\item{minheight}{Numeric value (default=4) from
\code{\link{classify_crown_pc}}.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
The projected crown area (numeric value) as the area of the concave
hull computed from the crown points of a tree point cloud.
}
\description{
Returns the projected crown area measured from a tree point cloud.
}
\details{
This function uses \code{\link[sf]{st_area}} and
\code{\link[concaveman]{concaveman}} to calculate the area of the concave
hull fitted to the crown points obtained with
\code{\link{classify_crown_pc}}.
}
\examples{
\dontrun{
PC_path <- "path/to/point_cloud.txt"
pc <- read_tree_pc(PC_path)
pca <- projected_crown_area_pc(pc)
pca <- projected_crown_area_pc(pc, 0.3)
pca <- projected_crown_area_pc(pc, 1, 1.5, 4, TRUE)
}
}
