% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavls_metrics_pc.R
\name{projected_tree_area_pc}
\alias{projected_tree_area_pc}
\title{Projected tree area point cloud}
\usage{
projected_tree_area_pc(pc, concavity = 2, plot = FALSE)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{concavity}{Numeric value (default=2) concavity for the computation of a
concave hull based on \code{\link[concaveman]{concaveman}}.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
The projected tree area (numeric value) as the area of the concave
hull computed from the tree points of a tree point cloud. Also optionally
(plot=TRUE) plots the concave hull fitting and in this case returns a list
with the pta as first element and the plot as the second element.
}
\description{
Returns the projected tree area measured from a tree point cloud.
}
\details{
This function uses \code{\link[sf]{st_area}} and
\code{\link[concaveman]{concaveman}} to calculate the area of the concave
hull fitted to the point cloud.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the projected crown area
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
pta <- projected_tree_area_pc(pc = pc_tree)
# and plot the concave hull fitting
output <- projected_tree_area_pc(pc = pc_tree, plot = TRUE)
pta <- output$pta
}
}
