% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree_pc.R
\name{read_tree_pc}
\alias{read_tree_pc}
\title{Read a tree point cloud}
\usage{
read_tree_pc(path, samplefactor = 1)
}
\arguments{
\item{path}{A character with the path to the tree point cloud file. File can
be \emph{txt}, \emph{las} or \emph{ply} format. The 3D coordinates have to
be in the first three columns in case of a \emph{txt} file.}

\item{samplefactor}{A numeric value ranging from 0 to 1 (default=1). This
determines the amount of points that are sampled from the point cloud. 1 to
sample 100 percent of the points and for example 0.5 to sample 50 percent
of the points.}
}
\value{
The tree point cloud as a data.frame (nx3) with X, Y, and Z
coordinates in the first, second and third column respectively.
}
\description{
Reads a tree point cloud file of txt, las or ply format and returns the tree
point cloud as a data.frame with 3 columns (X,Y,Z).
}
\details{
Reading the txt, las and ply files is based on
\code{\link[data.table]{fread}}, \code{\link[lidR]{readTLSLAS}} and
\code{\link[Rvcg]{vcgPlyRead}} respectively. Sampling is based on
\code{\link[base]{sample}} and is mainly a useful tool to reduce the amount
of points for quicker plotting.
}
\examples{
\dontrun{
# Read a tree point cloud file of the txt format
pc_txt <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
# Read a tree point cloud file of the ply format
pc_ply <- read_tree_pc(PC_path = "path/to/point_cloud.ply")
# Read a tree point cloud file of the las format
# and subsample to 20 percent of the points
pc_las <- read_tree_pc(PC_path = "path/to/point_cloud.las", 0.2)
}
}
