% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{shedding_ratio_qsm}
\alias{shedding_ratio_qsm}
\title{Calculate shedding ratio from a TreeQSM}
\usage{
shedding_ratio_qsm(branch, treedata)
}
\arguments{
\item{branch}{Branch field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}

\item{treedata}{Treedata field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The shedding ratio defined as "The number of stem branches without children
divided by the number of all branches in the bottom third" (Akerblom et al., 2017 & Terryn et al., 2020).
}
\description{
Calculate shedding ratio from a TreeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
shed_ratio <- shedding_ratio_qsm(branch, treedata)
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M. (2017). Automatic tree species recognition with quantitative structure models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ... & Verbeeck, H. (2020). Tree species classification using structural features derived from terrestrial laser scanning. ISPRS Journal of Photogrammetry and Remote Sensing, 168, 170-181.
}
