% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{stem_branch_cluster_size_qsm}
\alias{stem_branch_cluster_size_qsm}
\title{Calculate the stem branch cluster size from a TreeQSM}
\usage{
stem_branch_cluster_size_qsm(cylinder)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The stem branch cluster size defined as "the average number of 1st order
branches inside a 40cm height interval for 1st order branches. Each branch can
only belong to one interval" (Akerblom et al., 2017 & Terryn et al., 2020).
}
\description{
Calculate the stem branch cluster size from a TreeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
sbcs <- stem_branch_cluster_size_qsm(cylinder)
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M. (2017). Automatic tree species recognition with quantitative structure models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ... & Verbeeck, H. (2020). Tree species classification using structural features derived from terrestrial laser scanning. ISPRS Journal of Photogrammetry and Remote Sensing, 168, 170-181.
}
