% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{stem_branch_radius_qsm}
\alias{stem_branch_radius_qsm}
\title{Stem branch radius TreeQSM}
\usage{
stem_branch_radius_qsm(
  cylinder,
  treedata,
  normalisation = "treeheight",
  pc = NA
)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{normalisation}{Can be either "treeheight" or "parentcylinder". In case
of "treeheight" the mean radius of the 10 biggest branches is divided by
the tree height (Terryn et al., 2020). In case of "parentcylinder" the mean
is taken of the ratios of the radius of the 10 biggest branches and the
radius of their parent cylinders (Akerblom et al., 2017). When something
different than "treeheight" or "parentcylinder" is given, no normalisation
is done. Default is no normalisation.}

\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}. Default is NA and indicates no tree point cloud
is available. Only relevant if normalisation equals "treeheight".}
}
\value{
The stem branch radius. Unitless with normalisation, in meters
without normalisation.  NaN when there are no stem branches.
}
\description{
Calculates the stem branch radius from a TreeQSM.
}
\details{
The stem branch radius is defined as "Mean of the 10 largest 1st order
branches measured at the base. Can be normalised by the tree height or the
the stem radius at respective height" (Akerblom et al., 2017 & Terryn et al.,
2020). Tree height is calculated with \code{\link{tree_height}}.
}
\examples{
\dontrun{
# Read tree qsm and calculate the stem branch radius
# from Akerblom et al. (2017)
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
sbr <- stem_branch_radius_qsm(cylinder = qsm$cylinder,
                              treedata = qsm$treedata,
                              normalisation = "parentcyl")
# from Terryn et al. (2020)
sbr <- stem_branch_radius_qsm(cylinder = qsm$cylinder,
                              treedata = qsm$treedata,
                              normalisation = "treeheight")
# with point cloud data
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
sbr <- stem_branch_radius_qsm(cylinder = qsm$cylinder,
                              treedata = qsm$treedata,
                              normalisation = "treeheight",
                              pc = pc_tree)
}
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
