% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapering_metrics.R
\name{stem_tapering_pc}
\alias{stem_tapering_pc}
\title{Tapering of a trunk point cloud}
\usage{
stem_tapering_pc(
  pc,
  slice_thickness = 0.1,
  maxtaperheight = 10,
  interval = 1,
  buttress = FALSE,
  thresholdbuttress = 0.0015,
  maxbuttressheight = 5,
  plot = FALSE
)
}
\arguments{
\item{pc}{The trunk point cloud as a data.frame with columns X,Y,Z.}

\item{slice_thickness}{Numeric value (default = 0.1) that determines the
thickness of the slice which is used to measure the diameter at each
height.}

\item{maxtaperheight}{Numeric value (default = 10) that determines the
maximum height used for the tapering calculation.}

\item{interval}{Numeric value (default = 1) that determines the interval over
which the median of the diameters is taken to reduce influence of outliers
for the taper measurement. Choose interval value equal to the
slice_thickness if you want to use all the calculated diameters.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses (higher than breast height).}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant when buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter at breast
height. Only relevant when buttress == TRUE.}

\item{plot}{Logical (default=FALSE), indicates if the taper curve is plotted.}
}
\value{
list with tapering coefficients a and b, dataframe T with the
diameters at each respective height, and the taper plot.
}
\description{
Calculates the tapering of the trunk based on a linear regression on the
diameters at each height of the trunk.
}
\examples{
\dontrun{
# Read trunk point cloud
pc_trunk <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
# calculate the tapering
output <- stem_tapering_pc(pc_trunk)
}
}
