% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{summary_basic_pointcloud_metrics}
\alias{summary_basic_pointcloud_metrics}
\title{Summary basic structural metrics tree point cloud}
\usage{
summary_basic_pointcloud_metrics(
  PCs_path,
  extension = ".txt",
  thresholdbranch = 1.5,
  minheight = 1,
  concavity = 2,
  alpha = 1,
  buttress = FALSE,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7,
  OUT_path = FALSE
)
}
\arguments{
\item{PCs_path}{A character with the path to the folder that contains the
tree point clouds.}

\item{extension}{A character refering to the file extension of the point
cloud files (default=".txt"). Can be ".txt", ".ply" or ".las". Only
relevant if the tree point clouds are available.}

\item{thresholdbranch}{Numeric value (default=1.5) from
\code{\link{classify_crown_pc}}.}

\item{minheight}{Numeric value (default=1) from
\code{\link{classify_crown_pc}}. The default value is based on
non-buttressed trees. Choose a higher value (e.g. 4) for buttressed trees.}

\item{concavity}{Numeric value (default=2). Parameter of the
\code{\link{projected_crown_area_pc}} function used to calculate the
projected crown area.}

\item{alpha}{Numeric value (default=1). Parameter of the
\code{\link{volume_crown_pc}} function used to calculate the crown volume.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses (higher than breast height).}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant when buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant when buttress == TRUE.}

\item{OUT_path}{A character with the path to the folder where the summary csv
file should be saved or logical (default=FALSE) in this case no csv file is
produced.}
}
\value{
The summary of the basic structural metrics for multiple tree point
clouds as a data.frame. Includes the tree height, diameter at breast
height, diameter above buttresses, projected crown area and crown volume.
The summary is saved in a csv file if an output folder is provided.
}
\description{
Returns a summary data.frame containing the tree position (X,Y-coordinates),
tree height, diameter at breast height, diameter above buttresses, projected
crown area and crown volume.
}
\details{
The tree position, tree height, diameter at breast height, diameter above
buttresses, projected crown area and crown volume are otained with
\code{\link{tree_position_pc}}, \code{\link{tree_height_pc}},
\code{\link{dbh_pc}}, \code{\link{dab_pc}},
\code{\link{projected_crown_area_pc}} and \code{\link{volume_crown_pc}}
respectively.
}
\examples{
\dontrun{
# Calculate the summary with default parameters and export to csv
# recommended for non-buttressed trees
summary <- summary_basic_pointcloud_metrics(PCs_path = "path/to/folder/PCs/",
                                            OUT_path = "path/to/out/folder/")
# Calculate the summary with non-default parameter values
# recommended for buttressed trees
summary <- summary_basic_pointcloud_metrics(PCs_path = "path/to/folder/PCs/",
                                            extension = ".ply",
                                            minheight = 4, buttress = TRUE)
}
}
