% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{summary_basic_pointcloud_metrics_pertree}
\alias{summary_basic_pointcloud_metrics_pertree}
\title{Summary basic structural metrics tree point cloud for one tree.}
\usage{
summary_basic_pointcloud_metrics_pertree(
  PC_path,
  metrics = c("tree position", "stem diameter", "tree height", "projected area",
    "alpha volume"),
  dtm = NA,
  r = 5,
  crown = FALSE,
  thresholdbranch = 1.5,
  minheight = 1,
  concavity = 2,
  alpha = 1,
  buttress = FALSE,
  h_cutoff = 20,
  thresholdR2 = 0.001,
  slice_thickness = 0.06,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7,
  functional = TRUE,
  concavity_fdiameter = 4,
  OUT_path = FALSE,
  plot = FALSE,
  plotcolors = c("#000000", "#808080", "#1c027a", "#08aa7c", "#fac87f")
)
}
\arguments{
\item{PC_path}{A character with the path to the tree point cloud.}

\item{metrics}{A list of strings referring to the metrics you want to
calculate. The possibilities are "tree position", "stem diameter",
"tree height", "projected area", and "alpha volume".}

\item{dtm}{The digital terrain model from \code{\link{tree_height_pc}}.}

\item{r}{Numeric value (default=5) r which determines the range taken for the
dtm from \code{\link{tree_height_pc}}. Only relevant if a dtm is provided.}

\item{crown}{Logical (default=FALSE), indicates if the area and volume is
calculated based on the full point clouds (crown = FALSE) or only on the
crown point clouds (crown = TRUE).}

\item{thresholdbranch}{Numeric value (default=1.5) from
\code{\link{classify_crown_pc}}. Only relevant when crown == TRUE.}

\item{minheight}{Numeric value (default=1) from
\code{\link{classify_crown_pc}}. The default value is based on
non-buttressed trees. Choose a higher value (e.g. 4) for buttressed trees.
Only relevant when crown == TRUE.}

\item{concavity}{Numeric value (default=2). Parameter of the
\code{\link{projected_area_pc}} function used to calculate the projected
crown area.}

\item{alpha}{Numeric value (default=1). Parameter of the
\code{\link{alpha_volume_pc}} function used to calculate the crown volume.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses (higher than breast height).}

\item{h_cutoff}{Numeric value (default=15), height threshold where dab is
measured when buttress == TRUE.}

\item{thresholdR2}{Numeric value (default=0.001). Parameter of the
\code{\link{dbh_pc}} function used to calculate the diameter at breast
height. Only relevant if buttress == FALSE.}

\item{slice_thickness}{Numeric value (default = 0.06). Parameter of the
\code{\link{dbh_pc}} and \code{\link{dab_pc}} functions used to calculate
the diameter at breast height and above buttresses.}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant when buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant when buttress == TRUE.}

\item{functional}{Logical (default=FALSE), indicates if the functional
diameter should be calculated.}

\item{concavity_fdiameter}{Numeric value (default=4) concavity for the
computation of the functional diameter using a concave hull based on
\code{\link[concaveman]{concaveman}}. This concavity value is used in the
functions \code{\link{diameter_slice_pc}}, \code{\link{dbh_pc}},
\code{\link{dab_pc}}, and \code{\link{classify_crown_pc}}.}

\item{OUT_path}{A character with name of the output folder where the summary
figures should be saved or logical (default=FALSE) in this case no figures
are saved.}

\item{plot}{Logical (default=FALSE), indicates if summary figure for each
tree point cloud is plotted. If an OUT_path is provided, the figures are
saved in the OUT_path.}

\item{plotcolors}{list of five colors for plotting. Only relevant when plot =
TRUE. The stem points above buttresses, stem points at breast height,
fitted circle, the concave hull and the estimated center are colored by the
first, second, third, fourth and fifth element of this list respectively.}
}
\value{
The summary of the basic structural metrics for multiple tree point
clouds as a data.frame. Includes the tree height, diameter at breast
height, diameter above buttresses, projected (crown) area and (crown)
volume. The summary is saved in a csv file if an output folder is provided.
}
\description{
Returns a summary data.frame containing the tree position (X,Y-coordinates),
tree height, diameter at breast height, functional diameter at breast height,
diameter above buttresses, functional diameter above buttresses, projected
(crown) area and (crown) volume of one tree.
}
\details{
The tree position, tree height, diameter at breast height, functional
diameter at breast height, diameter above buttresses, functional diameter
above buttresses, projected (crown) area and (crown) volume are otained with
\code{\link{tree_position_pc}}, \code{\link{tree_height_pc}},
\code{\link{dbh_pc}}, \code{\link{dab_pc}}, \code{\link{projected_area_pc}}
and \code{\link{alpha_volume_pc}} respectively.
}
\examples{
\dontrun{
# Calculate the summary with default parameters and export summary figure
summary <- summary_basic_pointcloud_metrics(
  PC_path = "path/to/folder/PCs/",
  OUT_path = "path/to/out/folder/"
)
# Calculate the summary with non-default parameter values
# recommended for buttressed trees
summary <- summary_basic_pointcloud_metrics(
  PC_path = "path/to/folder/PCs/",
  metrics = c("tree position", "stem diameter"),
  crown = TRUE, minheight = 4, buttress = TRUE
)
}
}
