% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{summary_Terryn_2020}
\alias{summary_Terryn_2020}
\title{Summary features Terryn et al. (2020)}
\usage{
summary_Terryn_2020(
  QSMs_path,
  version = "2.4.0",
  PCs_path = "NA",
  buttress = FALSE,
  extension = ".txt"
)
}
\arguments{
\item{QSMs_path}{A charachter with the path to the folder that contains the
treeQSMs. These files have to be of the format xxx_000_qsm.mat (xxx is the
plotname, 000 is the tree number) or xxx_000_qsm_0.mat (0 at the end is for
example the n-th QSM that is made for tree 000). When multiple QSMs are
present for one tree the mean of the values of the different QSMs is taken
for that tree as a final value for a certain feature.}

\item{version}{A character indicating the version of TreeQSM that was used to
produce the qsms. Default version is 2.4.0. Other possible versions are
2.2.0.}

\item{PCs_path}{A charachter with the path to the folder that contains the
tree point clouds. Default is "NA" when the point clouds are not available.
The point clouds are used to determine the DBH, tree height, projected
crown area and crown volume. The DBH and tree height obtained from the tree
point clouds are then used for the normalisation of the other features. The
point cloud files have to be of the format xxx_000_pc in order to link the
tree point cloud to its' respective treeQSM.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses. Only relevant if the tree point clouds are available.}

\item{extension}{A character refering to the file extension of the point
cloud files (default=".txt"). Can be ".txt", ".ply" or ".las". Only
relevant if the tree point clouds are available.}
}
\value{
The summary of all metrics from Terryn et al. (2020) as a data.frame.
}
\description{
Returns a summary data.frame containing the tree height, dbh and the metrics
defined by Terryn et al. (2020). Also contains the projected crown area and
crown volume (NaN when point clouds are not available).
}
\details{
Metrics Terryn et al. (2020): stem branch angle (sba,
\code{\link{stem_branch_angle_qsm}}), stem branch cluster size (sbcs,
\code{\link{stem_branch_cluster_size_qsm}}), stem branch radius (sbr,
\code{\link{stem_branch_radius_qsm}}), stem branch length (sbl,
\code{\link{stem_branch_length_qsm}}), stem branch distance (sbd,
\code{\link{stem_branch_distance_qsm}}), dbh height ratio (dhr,
\code{\link{dbh_height_ratio_qsm}}), dbh volume ratio (dvr,
\code{\link{dbh_volume_ratio_qsm}}), volume below 55 (vb55,
\code{\link{volume_below_55_qsm}}), cylinder length volume ratio (clvr,
\code{\link{cylinder_length_volume_ratio_qsm}}), shedding ratio (sr,
\code{\link{shedding_ratio_qsm}}), branch angle ratio (bar,
\code{\link{branch_angle_ratio_qsm}}), relative volume ratio (rvr,
\code{\link{relative_volume_ratio_qsm}}), crown start height (csh,
\code{\link{crown_start_height_qsm}}), crown height (ch,
\code{\link{crown_height_qsm}}), crown evenness (ce,
\code{\link{crown_evenness_qsm}}), crown diameter height ratio (cdhr,
\code{\link{crown_diameterheight_ratio_qsm}}), dbh minimum radius ratio (dmr,
\code{\link{dbh_minradius_ratio_qsm}}). Tree height, DBH, crown projection
area and crown volume are determined with \code{\link{tree_height}},
\code{\link{dbh}}, \code{\link{projected_crown_area_pc}} and
\code{\link{volume_crown_pc}}. All functions are run with default parameters
(except for the parameters pc and buttress if point clouds are available.)
}
\examples{
\dontrun{
QSMs_path <- "path/to/folder/QSMs/"
PCs_path <- "path/to/folder/PCs/"
summary <- summary_Terryn_2020(QSMs_path)
summary <- summary_Terryn_2020(QSMs_path,version="2.4.0",PCs_path,
                               buttress=TRUE,extension=".txt")
}
}
\references{
Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y.,
Newnham, G., ... & Verbeeck, H. (2020). Tree species classification using
structural features derived from terrestrial laser scanning. ISPRS Journal
of Photogrammetry and Remote Sensing, 168, 170-181.
}
