% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{tree_height_pc}
\alias{tree_height_pc}
\title{Tree height point cloud}
\usage{
tree_height_pc(pc)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}
}
\value{
The tree height (numeric value).
}
\description{
Returns the tree height measured from a tree point cloud.
}
\details{
The tree height is measured as the difference between the Z-value of the
highest and lowest point of the tree point cloud.
}
\examples{
\dontrun{
PC_path <- "path/to/point_cloud.txt"
pc <- read_tree_pc(PC_path)
tree_height <- tree_height_pc(pc)
}
}
