% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{tree_position_pc}
\alias{tree_position_pc}
\title{Tree point cloud position}
\usage{
tree_position_pc(pc)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}
}
\value{
Numeric with the X, Y, Z coordinates (location) of the tree stem.
}
\description{
Returns the (X,Y,Z)-position of a tree point cloud based on the mean X, Y and
Z value of the 100 lowest points of the tree point cloud.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the tree position
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
pos <- tree_position_pc(pc = pc_tree)
}
}
