% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_structural_metrics.R
\name{tree_position_qsm}
\alias{tree_position_qsm}
\title{Determine the position of a treeQSM}
\usage{
tree_position_qsm(cylinder)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The XY coordinates (location) of the tree stem.
}
\description{
Determine the position of a treeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
pos <- tree_position_qsm(cylinder)
}
