% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{trunk_height_qsm}
\alias{trunk_height_qsm}
\title{Trunk height TreeQSM}
\usage{
trunk_height_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The trunk height of the TreeQSM in meters.
}
\description{
Extracts the trunk height from the treedata of a TreeQSM. For stems without
and with trangulation it selects the TrunkLenght and TriaTrunkLength object
from treedata respectively.
}
\examples{
\dontrun{
# Read tree qsm and extract tree height
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
trunk_height <- trunk_height_qsm(treedata = qsm$treedata)
}
}
