% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{trunk_length_qsm}
\alias{trunk_length_qsm}
\title{Trunk length QSM}
\usage{
trunk_length_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The trunk height of the TreeQSM in meters.
}
\description{
Extracts the trunk length from the treedata of a TreeQSM or RCT QSM. For
stems without and with trangulation (option in TreeQSM v2.4.x) it selects the
TrunkLenght and TriaTrunkLength object from treedata respectively.
}
\examples{
\dontrun{
# Read tree qsm and extract tree height
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
trunk_height <- trunk_height_qsm(treedata = qsm$treedata)
}
}
