% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{vertical_bin_radii}
\alias{vertical_bin_radii}
\title{Radii vertical bins TreeQSM}
\usage{
vertical_bin_radii(treedata, cylinder)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The radii of the three vertical bins.
}
\description{
Calculates the radii of the three vertical bins fitted to the TreeQSM
cylinders.
}
\details{
These radii are the radii of the cylinders whose axis are vertical and goes
through the bin centre point, and which contains approximately 90\% of the
volume of the branch cylinders in that bin. The tree is divided into three
vertical bins, and the centre point of each bin is defined as the average of
mean points of stem cylinders in the bin. If the bin does not contain stem
cylinders the centre of the previous bin is used (Akerblom et al., 2017 &
Terryn et al., 2020).
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
qsm <- read_tree_qsm(QSM_path)
radii <- vertical_bin_radii(qsm$treedata,qsm$cylinder)
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
