% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{volume_crown_pc}
\alias{volume_crown_pc}
\title{Crown volume point cloud}
\usage{
volume_crown_pc(
  pc,
  alpha = 1,
  thresholdbranch = 1.5,
  minheight = 4,
  buttress = FALSE,
  plot = FALSE
)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{alpha}{Numeric value (default=1) alpha for the computation of the 3D
alpha-shape of the tree crown based on
\code{\link[alphashape3d]{ashape3d}}.}

\item{thresholdbranch}{Numeric value (default=1.5) from
\code{\link{classify_crown_pc}}.}

\item{minheight}{Numeric value (default=4) from
\code{\link{classify_crown_pc}}.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses (higher than breast height). Only relevant if the tree point
clouds are available.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
The volume of the tree crown (numeric value) as the volume of the 3D
alpha-shape computed from the crown points of a tree point cloud.
}
\description{
Returns the crown volume measured from a tree point cloud.
}
\details{
This function uses \code{\link[alphashape3d]{ashape3d}} and
\code{\link[alphashape3d]{volume_ashape3d}} to calculate the volume of 3D the
alpha-shape fitted to the crown points obtained with
\code{\link{classify_crown_pc}}.
}
\examples{
\dontrun{
PC_path <- "path/to/point_cloud.txt"
pc <- read_tree_pc(PC_path, 1)
vol_crown <- volume_crown_pc(pc)
vol_crown <- volume_crown_pc(pc, 0.3, FALSE)
vol_crown <- volume_crown_pc(pc, 1, TRUE)
}
}
