% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavls_metrics_pc.R
\name{volume_tree_pc}
\alias{volume_tree_pc}
\title{Tree volume point cloud}
\usage{
volume_tree_pc(pc, alpha = 1, plot = FALSE)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{alpha}{Numeric value (default=1) alpha for the computation of the 3D
alpha-shape of the tree crown based on
\code{\link[alphashape3d]{ashape3d}}.}

\item{plot}{Logical (default=FALSE), indicates if the alpha-shape is plotted.}
}
\value{
The volume of the tree (numeric value) as the volume of the 3D
alpha-shape computed from the tree point cloud. Also optionally (plot=TRUE)
plots the alpha-shape and in this case returns a list with the point cloud
as first element and the alphashape3d object as the second element. The 3D
plot can be reconstructed using plot(output$alphashape3d).
}
\description{
Returns the tree volume measured from a tree point cloud.
}
\details{
This function uses \code{\link[alphashape3d]{ashape3d}} and
\code{\link[alphashape3d]{volume_ashape3d}} to calculate the volume of 3D the
alpha-shape fitted to the tree point cloud.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the crown volume
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
vol_tree <- volume_tree_pc(pc = pc_tree)
# and plot the 3D alpha-shape
output <- volume_tree_pc(pc = pc_tree, plot = TRUE)
vol_tree <- output$volume
}
}
