% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{classify_crown_pc}
\alias{classify_crown_pc}
\title{Crown classification}
\usage{
classify_crown_pc(pc, thresholdbranch = 1.5, minheight = 4, plot = FALSE)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{thresholdbranch}{Numeric value (default=1.5) that is multiplied with
the diameter of the tree (calculated with \code{\link{dab_pc}})
which determines the cutt-off where a branch emerges and the crown begins.}

\item{minheight}{Numeric value (default=4) with the minimum height at which
the crown begins. Should be above the widest part of the buttresses.}

\item{plot}{Logical (default=FALSE), indicates if the classified tree is
plotted.}
}
\value{
Data.frame with the crown point cloud (part of the tree above the
first branch). Also optionally (plot=TRUE) plots the crown vs non-crown
points.
}
\description{
Returns the crown points from a tree point cloud.
}
\details{
The classification is based on the increased distance between the minimum and
maximum X (and Y) coordinates of the tree points within a horizontal slice
when the first branch is reached with increasing height.
}
\examples{
PC_path <- "C:/Users/lmterryn/example_pointcloud2.txt"
pc <- read_tree_pc(PC_path,1)
crown_pc <- classify_crown_pc(pc,1.5,4,FALSE)
}
