% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{classify_crown_pc}
\alias{classify_crown_pc}
\title{Crown classification point cloud}
\usage{
classify_crown_pc(
  pc,
  thresholdbranch = 1.5,
  minheight = 1,
  buttress = FALSE,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7,
  plot = FALSE
)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{thresholdbranch}{Numeric value (default=1.5) that is multiplied with
the diameter of the tree (calculated with \code{\link{dbh_pc}} or
\code{\link{dab_pc}} when buttress =TRUE) which determines the cutt-off
where a branch emerges and the crown begins.}

\item{minheight}{Numeric value (default=1) with the minimum height at which
the crown begins. Should be above the widest part of the buttresses for
buttressed trees (value of 4 is recommended). For non-buttressed trees
choose a lower value (such as 1).}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses (higher than breast height).}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant when buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter at breast
height. Only relevant when buttress == TRUE.}

\item{plot}{Logical (default=FALSE), indicates if the classified tree is
plotted.}
}
\value{
Data.frame with the crown point cloud (part of the tree above the
first branch). Also optionally (plot=TRUE) plots the crown vs non-crown
points.
}
\description{
Returns the crown points from a tree point cloud.
}
\details{
The classification is based on the increased distance between the minimum and
maximum X (and Y) coordinates of the tree points within a horizontal slice
when the first branch is reached with increasing height.
}
\examples{
\dontrun{
# Read tree point cloud and extract the crown points
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
crown_pc <- classify_crown_pc(pc = pc_tree)
# and plot the classification results
crown_pc <- classify_crown_pc(pc = pc_tree, plot = TRUE)
# with non-default settings for a buttressed tree
crown_pc <- classify_crown_pc(pc = pc_tree, minheight = 4, buttress = TRUE,
                              plot = TRUE)
}
}
