% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{crown_diameterheight_ratio_qsm}
\alias{crown_diameterheight_ratio_qsm}
\title{Crown diameter height ratio TreeQSM}
\usage{
crown_diameterheight_ratio_qsm(treedata, cylinder, pc = "NA")
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}. Default is "NA" and indicates no tree point
cloud is available.}
}
\value{
The ratio of the crown diameter and crown height.
}
\description{
Calculates the ratio between the crown diameter and the crown height from a
TreeQSM (Akerblom et al., 2017 & Terryn et al., 2020).
}
\details{
The crown diameter is the maximum radii of the vertical bin radius estimates
that are calculated with \code{\link{vertical_bin_radii_qsm}}. The crown
height is the vertical distance between the highest and the lowest crown
cylinder and is obtained from \code{\link{crown_height_qsm}} multiplied with
the tree_height.
}
\examples{
\dontrun{
QSM_path <- "path/to/qsm.mat"
qsm <- read_tree_qsm(QSM_path)
cdh_ratio <- crown_diameterheight_ratio_qsm(qsm$treedata,qsm$cylinder,crown)
PC_path <- "path/to/point_cloud.txt"
pc <- read_tree_pc(PC_path)
cdh_ratio <- crown_diameterheight_ratio_qsm(qsm$treedata,qsm$cylinder,crown,
             pc)
}
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
