% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{cylinder_length_volume_ratio_qsm}
\alias{cylinder_length_volume_ratio_qsm}
\title{Cylinder length-volume ratio TreeQSM}
\usage{
cylinder_length_volume_ratio_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The cylinder length volume ratio in meters-2.
}
\description{
Calculates the cylinder length volume ratio from a TreeQSM.
}
\details{
The cylinder length volume ratio is defined as "the ratio between total
length and volume of the branch cylinders" (Akerblom et al., 2017 & Terryn et
al., 2020).
}
\examples{
\dontrun{
# Read tree qsm and calculate the cylinder length volume ratio
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
clvr <- cylinder_length_volume_ratio_qsm(treedata = qsm$treedata)
}
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
