% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{dbh}
\alias{dbh}
\title{Diameter at breast height/above buttresses}
\usage{
dbh(treedata, pc = NA, buttress = FALSE)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}. Default is NA and indicates no tree point cloud
is available.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses. Only relevant if pc is available.}
}
\value{
The dbh or dab in meters.
}
\description{
Returns the diameter at breast height or diameter above buttresses.
}
\details{
If the tree point cloud is available the calculations are based on the point
cloud (most accurate). In this case the diameter at breast height (dbh) or
diameter above buttresses (dab) is calculated with \code{\link{dbh_pc}} or
\code{\link{dab_pc}} respectively. If the tree point cloud is not available
the dbh is based on the treeQSM with \code{\link{dbh_qsm}}.
}
\examples{
\dontrun{
QSM_path <- "path/to/qsm.mat"
qsm <- read_tree_qsm(QSM_path)
DBH <- dbh(qsm)
PC_path <- "path/to/point_cloud.txt"
pc <- read_tree_pc(PC_path)
DBH <- dbh(qsm, pc, TRUE)
}
}
