% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{dbh_height_ratio_qsm}
\alias{dbh_height_ratio_qsm}
\title{DBH-tree height ratio TreeQSM}
\usage{
dbh_height_ratio_qsm(
  treedata,
  pc = NA,
  buttress = FALSE,
  thresholdR2 = 0.001,
  slice_thickness = 0.06,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7
)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}. Default is NA and indicates no tree point
cloud is available.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses. Only relevant if pc is available.}

\item{thresholdR2}{Numeric value (default=0.001). Parameter of the
\code{\link{dbh_pc}} function used to calculate the diameter at breast
height. Only relevant if the tree point cloud is available and buttress ==
FALSE.}

\item{slice_thickness}{Numeric value (default = 0.06). Parameter of the
\code{\link{dbh_pc}} function used to calculate the diameter at breast
height. Only relevant if the tree point cloud is available and buttress ==
FALSE.}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant if the tree point cloud is available and
buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter at breast
height. Only relevant if the tree point cloud is available and buttress
== TRUE.}
}
\value{
DBH divided by the tree height.
}
\description{
Calculates DBH-tree height ratio from a TreeQSM.
}
\details{
DBH and tree height are calculated with \code{\link{dbh}} and
\code{\link{tree_height}}.
}
\examples{
\dontrun{
# Read tree qsm and calculate the dbh height ratio
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
dhr <- dbh_height_ratio_qsm(treedata = qsm$treedata)
# with point cloud data
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
dhr <- dbh_height_ratio_qsm(treedata = qsm$treedata, pc = pc_tree)
# for buttressed trees
dhr <- dbh_height_ratio_qsm(
  treedata = qsm$treedata, pc = pc_tree,
  buttress = TRUE
)
}
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
