% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{dbh_minradius_ratio_qsm}
\alias{dbh_minradius_ratio_qsm}
\title{DBH minimum tree radius ratio TreeQSM}
\usage{
dbh_minradius_ratio_qsm(
  treedata,
  cylinder,
  pc = NA,
  buttress = FALSE,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7
)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}. Default is NA and indicates no tree point
cloud is available.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses. Only relevant if pc is}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant if the tree point cloud is available and
buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter at breast
height. Only relevant if the tree point cloud is available and buttress
== TRUE.}
}
\value{
The ratio of the dbh and the minimum tree radius.
}
\description{
Calculates the ratio between the dbh and the minimum tree radius from a
TreeQSM.
}
\details{
This ratio is defined as "Ratio between the DBH and the minimum of the
vertical bin radius estimates" (Akerblom et al., 2017 & Terryn et al., 2020).
The vertical bin radius estimates are calculated with
\code{\link{vertical_bin_radii_qsm}}. DBH is calculated with
\code{\link{dbh}}.
}
\examples{
\dontrun{
# Read tree qsm and calculate the dbh minimum tree radius ratio
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
dmrr <- dbh_minradius_ratio_qsm(treedata = qsm$treedata,
                                cylinder = qsm$cylinder)
# with point cloud data
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
dmrr <- dbh_minradius_ratio_qsm(treedata = qsm$treedata,
                                cylinder = qsm$cylinder, pc = pc_tree)
# for a buttressed tree
dmrr <- dbh_minradius_ratio_qsm(treedata = qsm$treedata,
                                cylinder = qsm$cylinder, pc = pc_tree,
                                buttress = TRUE)
}
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
