% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{dbh_pc}
\alias{dbh_pc}
\title{Diameter at breast height point cloud}
\usage{
dbh_pc(pc, thresholdR2 = 0.001, slice_thickness = 0.06, plot = FALSE)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{thresholdR2}{Numeric value (default=0.001) that is multiplied with the
radius to determine if at breast height (1.3 m above the lowest point of
the point cloud) the circle fit is influenced by branches. If the resulting
value is exceeded, the lower trunk without branches is extracted using
\code{\link{extract_lower_trunk_pc}}. Increase the thresholdR2 if your
point cloud quality is low (for example, errors in co-registration of point
clouds in multi-scan due to wind-effect).}

\item{slice_thickness}{Numeric value (default = 0.06) that determines the
thickness of the slice which is used to measure the diameter.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
Diameter of the stem at breast height (numeric value). Also
optionally (plot=TRUE) plots the circle fitting on the horizontal slice and
in this case returns a list with the dbh value as first element and the
plot as the second element.
}
\description{
Returns the diameter at breast height (DBH) of a tree measured from a tree
point cloud. There should be only one stem at breast height.
}
\details{
The DBH is measured as the diameter of the optimal circle fitted through a
6mm thick horizontal slice (from 1.27 m to 1.33 m above the lowest tree
point) using \code{\link{diameter_slice_pc}}. A least squares circle fitting
algorithm is applied to find the optimal fit. In case there are branches or
foliage at this height, the lower trunk is extracted using
\code{\link{extract_lower_trunk_pc}}. Wether this is the case is determined
using the thresholdR2 parameter.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the DBH
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
dbh <- dbh_pc(pc = pc_tree)
# and plot the circle fitting
output <- dbh_pc(pc = pc_tree, plot = TRUE)
dbh <- output$dbh
}
}
