% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{dbh_qsm}
\alias{dbh_qsm}
\title{Diameter at breast height TreeQSM}
\usage{
dbh_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The DBH of the TreeQSM in meters.
}
\description{
Extracts the dbh from the treedata of a TreeQSM.
}
\details{
The dbh is calculated as the diameter of the cylinder in the QSM at the right
height (cylinder at 1.3 m). If the trunk was modeled with triangulation the
dbh is calculated as mean length of the diagonals in the triangulation.
}
\examples{
\dontrun{
QSM_path <- "path/to/qsm.mat"
qsm <- read_tree_qsm(QSM_path)
dbh <- dbh_qsm(qsm$treedata)
}
}
