% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{DBH_qsm}
\alias{DBH_qsm}
\title{Diameter at breast height TreeQSM}
\usage{
DBH_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The DBH of the TreeQSM in meters.
}
\description{
Extracts the DBH from the treedata of a TreeQSM.
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
qsm <- read_tree_qsm(QSM_path)
dbh <- DBH_qsm(qsm$treedata)
}
