% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_structural_metrics.R
\name{DBH_qsm}
\alias{DBH_qsm}
\title{Extract the DBH from the treedata of a TreeQSM}
\usage{
DBH_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The DBH in m of the TreeQSM.
}
\description{
Extract the DBH from the treedata of a TreeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
dbh <- DBH_qsm(treedata)
}
