% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{normalize_pc}
\alias{normalize_pc}
\title{Normalize a tree point cloud}
\usage{
normalize_pc(pc)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}
}
\value{
Normalized point cloud as a data.frame with columns X,Y,Z.
}
\description{
Normalizes a tree point cloud by subtracting from each column its respective
the min value (e.g. all X-values - min(all X-values)).
}
\examples{
\dontrun{
PC_path <- "path/to/point_cloud.txt"
pc <- read_tree_pc(PC_path)
pc_norm <- normalize_pc(pc)
}
}
