% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{projected_crown_area_pc}
\alias{projected_crown_area_pc}
\title{Projected crown area point cloud}
\usage{
projected_crown_area_pc(
  pc,
  concavity = 2,
  thresholdbranch = 1.5,
  minheight = 1,
  buttress = FALSE,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7,
  plot = FALSE
)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{concavity}{Numeric value (default=2) concavity for the computation of a
concave hull based on \code{\link[concaveman]{concaveman}}.}

\item{thresholdbranch}{Numeric value (default=1.5) from
\code{\link{classify_crown_pc}}.}

\item{minheight}{Numeric value (default=1) from
\code{\link{classify_crown_pc}}.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses (higher than breast height).}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant when buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter at breast
height. Only relevant when buttress == TRUE.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
The projected crown area (numeric value) as the area of the concave
hull computed from the crown points of a tree point cloud.
}
\description{
Returns the projected crown area measured from a tree point cloud.
}
\details{
This function uses \code{\link[sf]{st_area}} and
\code{\link[concaveman]{concaveman}} to calculate the area of the concave
hull fitted to the crown points obtained with
\code{\link{classify_crown_pc}}.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the projected crown area
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
pca <- projected_crown_area_pc(pc = pc_tree)
# and plot the concave hull fitting
pca <- projected_crown_area_pc(pc = pc_tree, plot = TRUE)
# with non-default settings for a buttressed tree
pca <- projected_crown_area_pc(pc = pc_tree, concavity = 3, minheight = 4,
                               buttress = TRUE)
}
}
