% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{projected_crown_area_pc}
\alias{projected_crown_area_pc}
\title{Projected crown area}
\usage{
projected_crown_area_pc(pc, alpha, plot = FALSE)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{alpha}{Numeric value (default=1) alpha for the computation of a
alpha-convex hull based on \code{\link[alphahull]{ahull}}.}

\item{plot}{Logical (default=FALSE), indicates if the optimised circle
fitting is plotted.}
}
\value{
The projected crown area (numeric value) as the area of the
alpha-convex hull computed from the crown points of a tree point cloud.
}
\description{
Returns the projected crown area measured from a tree point cloud.
}
\details{
This function uses \code{\link[alphahull]{ahull}} to calculate the area of
the alpha-convex hull fitted to the crown points obtained with
\code{\link{classify_crown_pc}}.
}
\examples{
PC_path <- "C:/Users/lmterryn/example_pointcloud2.txt"
pc <- read_tree_pc(PC_path,0.1)
pca <- projected_crown_area_pc(pc,1,FALSE)
}
