% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree_qsm.R
\name{read_tree_qsm}
\alias{read_tree_qsm}
\title{Read a TreeQSM}
\usage{
read_tree_qsm(path, version = "2.4.0", global = FALSE)
}
\arguments{
\item{path}{The path to the TreeQSM mat file. This file contains 1 QSM
produced with \url{https://github.com/InverseTampere/TreeQSM} in matlab.}

\item{version}{A character indicating the version of TreeQSM that was used to
produce the TreeQSMs. Default version is "2.4.0" Other possible versions
are "2.0", "2.3.0", "2.3.1" and "2.3.2".}

\item{global}{Logical (default=FALSE), indicates if TreeQSM components should
be read into the global environment.}
}
\value{
Returns a list with the TreeQSM components (cylinder, branch,
treedata and triangulation) and optionally (global=TRUE) saves them into
the global environment.
}
\description{
Reads a TreeQSM matlab file (.mat) and returns its' components in a list and
optionally saves the TreeQSM components into the global environment.
}
\details{
Initial reading of the .mat file uses \code{\link[R.matlab]{readMat}}.
TreeQSM versions older than 2.4.0 are converted to the 2.4.0 structure, where
the value of attributes not found in the older version is set NA.
}
\examples{
\dontrun{
QSM_path <- "path/to/qsm.mat"
qsm <- read_tree_qsm(QSM_path)
qsm <- read_tree_qsm(QSM_path, version = "2.0")
qsm <- read_tree_qsm(QSM_path, version = "2.4.0", global = TRUE)
}
}
