% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree_qsm.R
\name{read_tree_qsm}
\alias{read_tree_qsm}
\title{Read a TreeQSM}
\usage{
read_tree_qsm(path)
}
\arguments{
\item{path}{The path to the TreeQSM mat file. This file contains 1 QSM
produced with version 2.4.0 of TreeQSM (https://github.com/InverseTampere/TreeQSM) in matlab.}
}
\value{
reads the QSM (tree) and its' components (qsm, cylinder, branch, treedata and triangulation) into the global environment.
}
\description{
Read a TreeQSM
}
\examples{
#QSM_path <- "path/to/test_qsm.txt"
#read_tree_qsm(QSM_path)
}
