% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{stem_branch_distance_qsm}
\alias{stem_branch_distance_qsm}
\title{Stem branch distance TreeQSM}
\usage{
stem_branch_distance_qsm(
  cylinder,
  treedata,
  normalisation = "no",
  pc = NA,
  buttress = FALSE,
  thresholdR2 = 0.001,
  slice_thickness = 0.06,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7
)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{normalisation}{Can either be "dbh" or nothing. In case of "dbh" the
average distance is divided by the DBH (Akerblom et al., 2017).}

\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}. Default is NA and indicates no tree point cloud
is available. Only relevant if normalisation equals "dbh".}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses. Only relevant if pc is available and normalisation equals
"dbh".}

\item{thresholdR2}{Numeric value (default=0.001). Parameter of the
\code{\link{dbh_pc}} function used to calculate the diameter at breast
height. Only relevant if the tree point cloud is available and buttress ==
FALSE.}

\item{slice_thickness}{Numeric value (default = 0.06). Parameter of the
\code{\link{dbh_pc}} and \code{\link{dab_pc}} functions used to calculate
the diameter at breast height and above buttresses. Only relevant if the
tree point cloud is available.}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant if the tree point cloud is available and buttress
== TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter at breast
height. Only relevant if the tree point cloud is available and buttress ==
TRUE.}
}
\value{
The stem branch distance. Unitless with normalisation, in meters
without normalisation. NaN when there are no stem branches.
}
\description{
Calculates the stem branch distance from a TreeQSM.
}
\details{
The stem branch distance is defined as "Average distance between 1st order
branches computed using a moving average with a window width 1 m. If window
is empty average distance in window is set as half of window width. Can be
normalised by the DBH" (Akerblom et al., 2017 & Terryn et al., 2020). When
something different than "dbh" is given, no normalisation is done. Default is
no normalisation. DBH is calculated with \code{\link{dbh}}.
}
\examples{
\dontrun{
# Read tree qsm and calculate the stem branch distance
# from Akerblom et al. (2017)
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
sbd <- stem_branch_distance_qsm(
  cylinder = qsm$cylinder,
  treedata = qsm$treedata,
  normalisation = "dbh"
)
# with point cloud data for buttressed trees
pc <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
sbd <- stem_branch_distance_qsm(
  cylinder = qsm$cylinder,
  treedata = qsm$treedata,
  normalisation = "dbh", pc = tree_pc,
  buttress = TRUE
)
# from Terryn et al. (2020)
sbd <- stem_branch_distance_qsm(
  cylinder = qsm$cylinder,
  treedata = qsm$treedata,
  normalisation = "no"
)
}
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
