% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Terryn_2020.R
\name{stem_branch_length_qsm}
\alias{stem_branch_length_qsm}
\title{Stem branch length TreeQSM}
\usage{
stem_branch_length_qsm(branch, treedata, normalisation = "no")
}
\arguments{
\item{branch}{Branch field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{normalisation}{Can be either "dbh" or "treeheight". In case of "dbh"
the mean of the lengths of the stem branches are divided by the DBH
(Akerblom et al., 2017). In case of "treeheight" the mean is divided by the
tree height (Terryn et al., 2020). When something different than "dbh" or
"treeheight" is given, no normalisation is done. Default is no
normalisation.}
}
\value{
The stem branch length.
}
\description{
Calculates the stem branch length from a TreeQSM.
}
\details{
The stem branch length is defined as "the average length of 1st order
branches. Can be normalised by DBH or tree height" (Akerblom et al., 2017 &
Terryn et al., 2020).
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
qsm <- read_tree_qsm(QSM_path)
sbl_dbh <- stem_branch_length_qsm(qsm$branch, qsm$treedata, "dbh")
sbl_th <- stem_branch_length_qsm(qsm$branch, qsm$treedata, "treeheight")
sbl_nn <- stem_branch_length_qsm(qsm$branch, qsm$treedata, "no")
}
\references{
Akerblom, M., Raumonen, P., Makipaa, R., & Kaasalainen, M.
(2017). Automatic tree species recognition with quantitative structure
models. Remote Sensing of Environment, 191, 1-12.

Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y., Newnham, G., ...
& Verbeeck, H. (2020). Tree species classification using structural
features derived from terrestrial laser scanning. ISPRS Journal of
Photogrammetry and Remote Sensing, 168, 170-181.
}
