% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{summary_basic_pointcloud_metrics}
\alias{summary_basic_pointcloud_metrics}
\title{Summary basic structural metrics tree point cloud}
\usage{
summary_basic_pointcloud_metrics(
  PCs_path,
  extension = ".txt",
  thresholdbuttress = 0.001,
  maxbuttressheight = 9,
  concavity = 2,
  alpha = 1
)
}
\arguments{
\item{PCs_path}{A character with the path to the folder that contains the
tree point clouds.}

\item{extension}{A character refering to the file extension of the point
cloud files (default=".txt"). Can be ".txt", ".ply" or ".las". Only
relevant if the tree point clouds are available.}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses.}

\item{maxbuttressheight}{Numeric value (default=9). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter at breast
height.}

\item{concavity}{Numeric value (default=2). Parameter of the
\code{\link{projected_crown_area_pc}} function used to calculate the
projected crown area.}

\item{alpha}{Numeric value (default=1). Parameter of the
\code{\link{volume_crown_pc}} function used to calculate the crown volume.}
}
\value{
The summary of the basic structural metrics for multiple tree point
clouds as a data.frame. Includes the tree height, diameter at breast
height, diameter above buttresses, projected crown area and crown volume.
}
\description{
Returns a summary data.frame containing the tree position (X,Y-coordinates),
tree height, diameter at breast height, diameter above buttresses, projected
crown area and crown volume.
}
\details{
The tree position, tree height, diameter at breast height, diameter above
buttresses, projected crown area and crown volume are otained with
\code{\link{tree_position_pc}}, \code{\link{tree_height_pc}},
\code{\link{dbh_pc}}, \code{\link{dab_pc}},
\code{\link{projected_crown_area_pc}} and \code{\link{volume_crown_pc}}
respectively.
}
\examples{
\dontrun{
PCs_path <- "path/to/folder/PCs/"
summary <- summary_basic_pointcloud_metrics(PCs_path)
summary <- summary_basic_pointcloud_metrics(PCs_path,extension=".ply")
}
}
