% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{summary_Terryn_2020}
\alias{summary_Terryn_2020}
\title{Summary structural metrics Terryn et al. (2020)}
\usage{
summary_Terryn_2020(
  QSMs_path,
  version = "2.4.0",
  sbr_normalisation = "treeheight",
  sbl_normalisation = "treeheight",
  sbd_normalisation = "no",
  PCs_path = NA,
  extension = ".txt",
  buttress = FALSE,
  thresholdR2 = 0.001,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7,
  OUT_path = FALSE
)
}
\arguments{
\item{QSMs_path}{A character with the path to the folder that contains the
treeQSMs. These files have to be of the format xxx_000_qsm.mat (xxx is the
plotname, 000 is the tree number) or xxx_000_qsm_0.mat (0 at the end is for
example the n-th QSM that is made for tree 000). When multiple QSMs are
present for one tree the mean of the values of the different QSMs is taken
for that tree as a final value for a certain feature.}

\item{version}{A character indicating the version of TreeQSM that was used to
produce the qsms. Default version is 2.4.0. Other possible versions are
2.2.0.}

\item{sbr_normalisation}{Character (default="treeheight"). Normalisation
parameter of \code{\link{stem_branch_radius_qsm}}.}

\item{sbl_normalisation}{Character (default="treeheight"). Normalisation
parameter of \code{\link{stem_branch_length_qsm}}.}

\item{sbd_normalisation}{Character (default="no"). Normalisation parameter of
\code{\link{stem_branch_distance_qsm}}.}

\item{PCs_path}{A character with the path to the folder that contains the
tree point clouds. Default is NA when the point clouds are not available.
The point clouds are used to determine the DBH, tree height, projected
crown area and crown volume. The DBH and tree height obtained from the tree
point clouds are then used for the normalisation of the other features. The
point cloud files have to be of the format xxx_000_pc in order to link the
tree point cloud to its' respective treeQSM.}

\item{extension}{A character refering to the file extension of the point
cloud files (default=".txt"). Can be ".txt", ".ply" or ".las". Only
relevant if the tree point clouds are available.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses. Only relevant if the tree point clouds are available. Only
relevant if the tree point clouds are available.}

\item{thresholdR2}{Numeric value (default=0.001). Parameter of the
\code{\link{dbh_pc}} function used to calculate the diameter at breast
height. Only relevant if the tree point cloud is available and buttress ==
FALSE.}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant if the tree point clouds are available and
buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant if the tree point clouds are available and
buttress == TRUE.}

\item{OUT_path}{A character with the path to the folder where the summary csv
file should be saved or logical (default=FALSE) in this case no csv file is
produced.}
}
\value{
The summary of all metrics from Terryn et al. (2020) as a data.frame.
The summary is saved in a csv file if an output folder is provided. If
multiple QSMs are provided for all trees the mean values and standard
deviations for each tree are also calculated and saved in 2 other csv
files. In this case the function returns a list of the summaries with the
means and standard deviations in the second and third element of the list
respectively.
}
\description{
Returns a summary data.frame containing all the metrics defined by Terryn et
al. (2020). Also contains: X and Y-position, dbh, tree height, tree volume
and trunk volume. When tree point clouds are provided, dbh and tree height
are based on the point cloud instead of the QSM.
}
\details{
Metrics Terryn et al. (2020): stem branch angle (sba,
\code{\link{stem_branch_angle_qsm}}), stem branch cluster size (sbcs,
\code{\link{stem_branch_cluster_size_qsm}}), stem branch radius (sbr,
\code{\link{stem_branch_radius_qsm}}), stem branch length (sbl,
\code{\link{stem_branch_length_qsm}}), stem branch distance (sbd,
\code{\link{stem_branch_distance_qsm}}), dbh height ratio (dhr,
\code{\link{dbh_height_ratio_qsm}}), dbh volume ratio (dvr,
\code{\link{dbh_volume_ratio_qsm}}), volume below 55 (vb55,
\code{\link{volume_below_55_qsm}}), cylinder length volume ratio (clvr,
\code{\link{cylinder_length_volume_ratio_qsm}}), shedding ratio (sr,
\code{\link{shedding_ratio_qsm}}), branch angle ratio (bar,
\code{\link{branch_angle_ratio_qsm}}), relative volume ratio (rvr,
\code{\link{relative_volume_ratio_qsm}}), crown start height (csh,
\code{\link{crown_start_height_qsm}}), crown height (ch,
\code{\link{crown_height_qsm}}), crown evenness (ce,
\code{\link{crown_evenness_qsm}}), crown diameter height ratio (cdhr,
\code{\link{crown_diameterheight_ratio_qsm}}), dbh minimum radius ratio (dmr,
\code{\link{dbh_minradius_ratio_qsm}}).
}
\examples{
\dontrun{
# Calculate the summary with default parameters and export to csv
# recommended for non-buttressed trees
summary <- summary_Terryn_2020(QSMs_path = "path/to/folder/QSMs/",
                               OUT_path = "path/to/out/folder/")
# also using point cloud info
summary <- summary_Terryn_2020(QSMs_path = "path/to/folder/QSMs/",
                               PCs_path = "path/to/folder/PCs/",
                               extension = ".txt",
                               OUT_path = "path/to/out/folder/")
# Calculate the summary with non-default parameter values
# recommended for buttressed trees
summary <- summary_Terryn_2020(QSMs_path = "path/to/folder/QSMs/",
                               PCs_path = "path/to/folder/PCs/",
                               extension = ".txt", buttress = TRUE,
                               OUT_path = "path/to/out/folder/")
}
}
\references{
Terryn, L., Calders, K., Disney, M., Origo, N., Malhi, Y.,
Newnham, G., ... & Verbeeck, H. (2020). Tree species classification using
structural features derived from terrestrial laser scanning. ISPRS Journal
of Photogrammetry and Remote Sensing, 168, 170-181.
}
