% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{total_branch_volume_qsm}
\alias{total_branch_volume_qsm}
\title{Total branch volume TreeQSM}
\usage{
total_branch_volume_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The total branch volume of the TreeQSM in liters.
}
\description{
Extracts the total branch volume from the treedata of a TreeQSM.
}
\examples{
\dontrun{
# Read tree qsm and extract total branch volume
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
branchvol <- total_branch_volume_qsm(treedata = qsm$treedata)
}
}
