% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_structural_metrics.R
\name{total_branch_volume_qsm}
\alias{total_branch_volume_qsm}
\title{Extract the total branch volume from the treedata of a TreeQSM}
\usage{
total_branch_volume_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The total branch volume in L of the TreeQSM.
}
\description{
Extract the total branch volume from the treedata of a TreeQSM
}
\examples{
#QSM_path <- "path/to/test_qsm.txt"
#read_tree_qsm(QSM_path)
#branchvol <- total_branch_volume_qsm(treedata)
}
