% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_structural_metrics.R
\name{total_cyl_length_qsm}
\alias{total_cyl_length_qsm}
\title{Extract the total cylinder length from the treedata of a TreeQSM}
\usage{
total_cyl_length_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The total length of all the cylinders (branch and trunk) in m of a TreeQSM.
}
\description{
Extract the total cylinder length from the treedata of a TreeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
cyllen <- total_cyl_length_qsm(treedata)
}
