% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_structural_metrics.R
\name{tree_height_qsm}
\alias{tree_height_qsm}
\title{Extract the tree height from the treedata of a TreeQSM}
\usage{
tree_height_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The tree height in m of the TreeQSM.
}
\description{
Extract the tree height from the treedata of a TreeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
height <- tree_height_qsm(treedata)
}
