% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{tree_height_qsm}
\alias{tree_height_qsm}
\title{Tree height TreeQSM}
\usage{
tree_height_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The tree height of the TreeQSM in meters.
}
\description{
Extracts the tree height from the treedata of a TreeQSM.
}
\examples{
\dontrun{
# Read tree qsm and extract tree height
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
height <- tree_height_qsm(treedata = qsm$treedata)
}
}
