% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{tree_position_pc}
\alias{tree_position_pc}
\title{Tree point cloud position}
\usage{
tree_position_pc(pc)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}
}
\value{
Numeric with the XY coordinates (location) of the tree stem.
}
\description{
Returns the (X,Y)-position of a tree point cloud based on the mean X and Y
value of the points within a 6mm thick horizontal slice at breast height
(from 1.27 m to 1.33 m above the lowest tree point).
}
\examples{
\dontrun{
PC_path <- "path/to/point_cloud.txt"
pc <- read_tree_pc(PC_path)
pos <- tree_position_pc(pc)
}
}
