% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{tree_position_qsm}
\alias{tree_position_qsm}
\title{TreeQSM position}
\usage{
tree_position_qsm(cylinder)
}
\arguments{
\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
Numeric with the XY coordinates (location) of the tree stem.
}
\description{
Returns the (X,Y)-position of a treeQSM based on the start position of the
first cylinder that is higher than 1.3 m above ground.
}
\examples{
\dontrun{
QSM_path <- "path/to/qsm.mat"
qsm <- read_tree_qsm(QSM_path)
pos <- tree_position_qsm(qsm$cylinder)
}
}
