% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{tree_volume_qsm}
\alias{tree_volume_qsm}
\title{Total tree volume TreeQSM}
\usage{
tree_volume_qsm(treedata, cylinder = NA, cylindercutoff = 0)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}

\item{cylinder}{Cylinder field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}, (default = NA - you do not need cylinder when
cylindercutoff == 0).}

\item{cylindercutoff}{This is the cutoff radius in meters for which cylinders
are to be included in the volume calculation.Default of 0 includes all
cylinders.}
}
\value{
The total volume of the TreeQSM in liters. If the trunk was modeled
with triangulation the total volume is the sum of the triangulated volume
of the stem (bottom), the volume of the stem cylinder (top) and the volume
of the branch cylinders.
}
\description{
Extracts the total tree volume from the treedata of a TreeQSM.
}
\examples{
\dontrun{
# Read tree qsm and extract tree volume
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
tot_vol <- tree_volume_qsm(treedata = qsm$treedata)
# Only include cylinders larger than 2.5 cm in radius
tot_vol <- tree_volume_qsm(treedata = qsm$treedata, cylinder = qsm$cylinder,
                           cylindercutoff = 0.25)
}
}
