% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_structural_metrics.R
\name{tree_volume_qsm}
\alias{tree_volume_qsm}
\title{Extract the total tree volume from the treedata of a TreeQSM}
\usage{
tree_volume_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM which is imported in the
global environment with the read_tree_qsm function.}
}
\value{
The total volume in L of the TreeQSM. If the trunk was modelled with triangulation
the total volume is the sum of the triangulated volume of the stem (bottom),
the volume of the stem cylinder (top) and the volume of the branch cylinders.
}
\description{
Extract the total tree volume from the treedata of a TreeQSM
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
read_tree_qsm(QSM_path)
totvol <- tree_volume_qsm(treedata)
}
