% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_qsm.R
\name{tree_volume_qsm}
\alias{tree_volume_qsm}
\title{Total tree volume TreeQSM}
\usage{
tree_volume_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The total volume of the TreeQSM in liters. If the trunk was modeled
with triangulation the total volume is the sum of the triangulated volume
of the stem (bottom), the volume of the stem cylinder (top) and the volume
of the branch cylinders.
}
\description{
Extracts the total tree volume from the treedata of a TreeQSM.
}
\examples{
\dontrun{
# Read tree qsm and extract tree volume
qsm <- read_tree_qsm(QSM_path = "path/to/qsm.mat")
tot_vol <- tree_volume_qsm(treedata = qsm$treedata)
}
}
