% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_structural_metrics.R
\name{trunk_volume_qsm}
\alias{trunk_volume_qsm}
\title{Total trunk volume TreeQSM}
\usage{
trunk_volume_qsm(treedata)
}
\arguments{
\item{treedata}{Treedata field of a TreeQSM that is returned by
\code{\link{read_tree_qsm}}.}
}
\value{
The total trunk volume of the TreeQSM in liters. If the trunk was
modelled with triangulation the total volume is the sum of the triangulated
volume of the stem (bottom) and the volume of the stem cylinder (top).
}
\description{
Extracts the total trunk volume from the treedata of a TreeQSM.
}
\examples{
QSM_path <- "C:/Users/lmterryn/example_qsm.mat"
qsm <- read_tree_qsm(QSM_path)
trunkvol <- trunk_volume_qsm(qsm$treedata)
}
