% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_metrics_pc.R
\name{volume_crown_pc}
\alias{volume_crown_pc}
\title{Crown volume point cloud}
\usage{
volume_crown_pc(
  pc,
  alpha = 1,
  thresholdbranch = 1.5,
  minheight = 1,
  buttress = FALSE,
  thresholdbuttress = 0.001,
  maxbuttressheight = 7,
  plot = FALSE
)
}
\arguments{
\item{pc}{The tree point cloud as a data.frame with columns X,Y,Z. Output of
\code{\link{read_tree_pc}}.}

\item{alpha}{Numeric value (default=1) alpha for the computation of the 3D
alpha-shape of the tree crown based on
\code{\link[alphashape3d]{ashape3d}}.}

\item{thresholdbranch}{Numeric value (default=1.5) from
\code{\link{classify_crown_pc}}.}

\item{minheight}{Numeric value (default=1) from
\code{\link{classify_crown_pc}}.}

\item{buttress}{Logical (default=FALSE), indicates if the trees have
buttresses (higher than breast height). Only relevant if the tree point
clouds are available.}

\item{thresholdbuttress}{Numeric value (default=0.001). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter above
buttresses. Only relevant when buttress == TRUE.}

\item{maxbuttressheight}{Numeric value (default=7). Parameter of the
\code{\link{dab_pc}} function used to calculate the diameter at breast
height. Only relevant when buttress == TRUE.}

\item{plot}{Logical (default=FALSE), indicates if the alpha-shape is plotted.}
}
\value{
The volume of the tree crown (numeric value) as the volume of the 3D
alpha-shape computed from the crown points of a tree point cloud. Also
optionally (plot=TRUE) plots the alpha-shape and in this case returns a
list with the crown point cloud as first element and the alphashape3d
object as the second element. The 3D plot can be reconstructed using
plot(output$alphashape3d).
}
\description{
Returns the crown volume measured from a tree point cloud.
}
\details{
This function uses \code{\link[alphashape3d]{ashape3d}} and
\code{\link[alphashape3d]{volume_ashape3d}} to calculate the volume of 3D the
alpha-shape fitted to the crown points obtained with
\code{\link{classify_crown_pc}}.
}
\examples{
\dontrun{
# Read tree point cloud and calculate the crown volume
pc_tree <- read_tree_pc(PC_path = "path/to/point_cloud.txt")
vol_crown <- volume_crown_pc(pc = pc_tree)
# and plot the 3D alpha-shape
output <- volume_crown_pc(pc = pc_tree, plot = TRUE)
vol_crown <- output$volume
# with non-default settings for a buttressed tree
vol_crown <- volume_crown_pc(pc = pc_tree, alpha = 2, minheight = 4)
}
}
